requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-groq>=0.1.6
langchain-community>=0.2.0
langchain-core>=0.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
tavily-python>=0.3.3
pypdf>=4.2.0
python-dotenv>=1.0.1
protobuf>=4.25.0
